Remove use of sprintf after running Psalm on the code base
While I don't see an issue with sprintf, I love using it, it seems that
direct interpolation in the call to fwrite is a better choice, due to
the previous call using %s which expects values to be a bool, int, or a
string, I believe. While I was making that change, I changed the sprintf
calls in the InvalidArgumentExceptions as well, for consistency.